restore preprocessing dict
